requirements.txt
Flask
flask-sqlalchemy
Python-dotenv
virtualenv
psycopg2-binary